Combine wsFetcher and httpFetcher with universalFetcher
